# specs/openapi.yaml
openapi: 3.1.0
info:
  title: Seatsurfing REST API
  description: |
    Seatsurfing provides REST APIs used by the mobile and web app as well as the administrator web interface.

    ## Authorization checks

    Each API endpoint is protected by authorization checks:
    1. **Role check**: Role of the user (Anonymous, User, Space Admin, Org Admin, Super Admin)
    2. **Organization check**: Organization membership

    ## Roles

    - **Anonymous**: Every user allowed, no authorization required
    - **User**: Authorized user who is a member of an organization
    - **Space Admin**: An organization's space administrator (can manage locations, spaces, bookings)
    - **Org Admin**: An organization's full administrator
    - **Super Admin**: Global administrator with access to all organizations and all functions

    ## Authentication

    Authentication is done via JWT Bearer tokens in the `Authorization` header. Service accounts may also use HTTP Basic Authentication.

    ## Session management

    - Access Token Lifetime: 15 minutes
    - Refresh Token Lifetime: 28 days
    - Tokens are signed using RS512

    ## Custom error codes

    Some endpoints return custom error codes via the `X-Error-Code` response header:
    - `1001`: Booking slot conflict
    - `1002`: Location max concurrent bookings reached
    - `1003`: Too many upcoming bookings for user
    - `1004`: Too many days in advance
    - `1005`: Invalid booking duration
...
    - `1010`: Booking subject required
...
servers:
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token obtained via login or refresh endpoints
    BasicAuth:
      type: http
      scheme: basic
      description: Service account authentication using `{orgId}_{email}` as username and the service account password
  schemas:
...
    CreateBookingRequest:
      type: object
      required: [spaceId, enter, leave]
      properties:
        spaceId:
          type: string
          format: uuid
        enter:
          type: string
          format: date-time
        leave:
          type: string
          format: date-time
        subject:
          type: string
        userEmail:
          type: string
          format: email
          description: Email of the user to book for (Space Admin only)
    GetBookingResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        userEmail:
          type: string
          format: email
        approved:
          type: boolean
        spaceId:
          type: string
          format: uuid
        subject:
          type: string
        enter:
          type: string
          format: date-time
        leave:
          type: string
          format: date-time
        recurringId:
          type: string
        space:
          $ref: "#/components/schemas/GetSpaceResponse"
    PreCreateBookingRequest:
      type: object
      required: [locationId, enter, leave]
      properties:
        locationId:
          type: string
          format: uuid
        enter:
          type: string
          format: date-time
        leave:
          type: string
          format: date-time
...
    CreateRecurringBookingRequest:
      type: object
      required: [spaceId, enter, leave, end, cadence, cycle]
      properties:
        spaceId:
          type: string
          format: uuid
        subject:
          type: string
        enter:
          type: string
          format: date-time
          description: Start time of the first booking
        leave:
          type: string
          format: date-time
          description: End time of the first booking
        end:
          type: string
          format: date-time
          description: End date of the recurrence
        cadence:
          type: integer
          minimum: 1
          maximum: 2
          description: "1 = weekly, 2 = daily"
        cycle:
          type: integer
          minimum: 1
          description: "Repeat every N weeks/days"
        weekdays:
          type: array
          items:
            type: integer
            minimum: 0
            maximum: 6
          description: "Weekdays (0=Sunday..6=Saturday), used with weekly cadence"
    CreateRecurringBookingResponse:
      type: object
      properties:
        enter:
          type: string
          format: date-time
        leave:
          type: string
          format: date-time
        success:
          type: boolean
        errorCode:
          type: integer
        id:
          type: string
          format: uuid
    GetRecurringBookingResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        spaceId:
          type: string
          format: uuid
        subject:
          type: string
        enter:
          type: string
          format: date-time
        leave:
          type: string
          format: date-time
        end:
          type: string
          format: date-time
        cadence:
          type: integer
        cycle:
          type: integer
        weekdays:
          type: array
          items:
            type: integer

    # --- Locations ---
...
    GetLocationResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        organizationId:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
        maxConcurrentBookings:
          type: integer
        timezone:
          type: string
        enabled:
          type: boolean
        mapScale:
          type: number
          format: double
        mapWidth:
          type: integer
        mapHeight:
          type: integer
        mapMimeType:
          type: string
...
    SpaceAttributeValueRequest:
      type: object
      properties:
        attributeId:
          type: string
          format: uuid
        value:
          type: string
    GetSpaceResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        locationId:
          type: string
          format: uuid
        available:
          type: boolean
        name:
          type: string
        x:
          type: integer
        "y":
          type: integer
        width:
          type: integer
        height:
          type: integer
        rotation:
          type: integer
        requireSubject:
          type: boolean
        enabled:
          type: boolean
        spaceTypeId:
          type: string
          format: uuid
        spaceType:
          $ref: "#/components/schemas/GetSpaceTypeResponse"
        attributes:
          type: array
          items:
            $ref: "#/components/schemas/SpaceAttributeValueRequest"
        approverGroupIds:
          type: array
          items:
            type: string
        allowedBookerGroupIds:
          type: array
          items:
            type: string
        location:
          $ref: "#/components/schemas/GetLocationResponse"
...
    CreateSpaceTypeRequest:
      type: object
      required: [name, bookingMode]
      properties:
        name:
          type: string
          maxLength: 128
        bookingMode:
          type: string
          enum: [flexible_time, fixed_slots]
        minDurationMinutes:
          type: integer
          minimum: 0
        enabled:
          type: boolean
    CreateSpaceTypeSlotRequest:
      type: object
      required: [label, startTime, endTime]
      properties:
        label:
          type: string
          maxLength: 128
        startTime:
          type: string
          pattern: "^\\d{2}:\\d{2}$"
          example: "08:00"
        endTime:
          type: string
          pattern: "^\\d{2}:\\d{2}$"
          example: "12:00"
        enabled:
          type: boolean
        sortOrder:
          type: integer
    GetSpaceTypeSlotResponse:
      allOf:
        - $ref: "#/components/schemas/CreateSpaceTypeSlotRequest"
        - type: object
          properties:
            id:
              type: string
              format: uuid
            spaceTypeId:
              type: string
              format: uuid
    GetSpaceTypeResponse:
      allOf:
        - $ref: "#/components/schemas/CreateSpaceTypeRequest"
        - type: object
          properties:
            id:
              type: string
              format: uuid
            organizationId:
              type: string
              format: uuid
            slots:
              type: array
              items:
                $ref: "#/components/schemas/GetSpaceTypeSlotResponse"
...
  responses:
    Created:
      description: Object created successfully
      headers:
        X-Object-ID:
          schema:
            type: string
          description: ID of the created object
    Updated:
      description: Object updated or deleted successfully (204 No Content)
    BadRequest:
      description: Malformed request or incomplete JSON
      headers:
        X-Error-Code:
          schema:
            type: integer
          description: Custom error code (see error codes table)
...
    Forbidden:
      description: Authorized, but user has no access to requested resource
    NotFound:
      description: Object not found
    Conflict:
      description: Conflicting request (e.g. slot already booked)
    PaymentRequired:
      description: Feature not available (license/plan restriction)
    InternalServerError:
      description: Internal server error
paths:
...
  /booking/:
    get:
      tags: [Bookings]
      summary: Get own upcoming bookings
      description: Returns the authenticated user's upcoming bookings (bookings that haven't ended yet in the location's timezone).
      operationId: getOwnBookings
      security:
        - BearerAuth: []
      responses:
        "200":
          description: List of upcoming bookings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GetBookingResponse"
        "500":
          $ref: "#/components/responses/InternalServerError"
    post:
      tags: [Bookings]
      summary: Create a booking
      description: |
        Creates a new booking. Space Admins can book on behalf of other users by providing `userEmail`.
        Various validation checks are performed (max bookings, max days in advance, slot conflicts, etc.).
      operationId: createBooking
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBookingRequest"
      responses:
        "201":
          $ref: "#/components/responses/Created"
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          description: Slot conflict (space already booked for that time)
        "500":
          $ref: "#/components/responses/InternalServerError"
  /booking/{id}:
    get:
      tags: [Bookings]
      summary: Get a booking
      description: Returns a single booking by ID. Users can view their own bookings; Space Admins can view any booking in their organization.
      operationId: getBooking
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Booking details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GetBookingResponse"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
    put:
      tags: [Bookings]
      summary: Update a booking
      description: Updates an existing booking. Users can update their own bookings; Space Admins can update any booking.
      operationId: updateBooking
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBookingRequest"
      responses:
        "204":
          $ref: "#/components/responses/Updated"
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
    delete:
      tags: [Bookings]
      summary: Delete a booking
      description: Deletes a booking. Users can delete their own bookings; Space Admins can delete any booking. Respects the "max hours before delete" setting.
      operationId: deleteBooking
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "204":
          $ref: "#/components/responses/Updated"
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /booking/{id}/ical:
    get:
      tags: [Bookings]
      summary: Download booking as iCal
      description: Downloads the booking as an iCalendar (.ics) file.
      operationId: getBookingIcal
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: iCalendar file
          content:
            text/calendar:
              schema:
                type: string
          headers:
            Content-Disposition:
              schema:
                type: string
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /booking/precheck/:
    post:
      tags: [Bookings]
      summary: Pre-check booking creation
      description: Validates whether a booking can be created for the given time range and location without actually creating it. Useful for client-side validation.
      operationId: preBookingCreateCheck
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PreCreateBookingRequest"
      responses:
        "204":
          description: Booking would be valid
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
...
  /recurring-booking/:
    post:
      tags: [Recurring Bookings]
      summary: Create a recurring booking
      description: Creates a recurring booking series. Individual bookings are created for each occurrence. Returns the result for each occurrence indicating success or failure.
      operationId: createRecurringBooking
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRecurringBookingRequest"
      responses:
        "201":
          description: Recurring booking created
          headers:
            X-Object-ID:
              schema:
                type: string
              description: ID of the recurring booking
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CreateRecurringBookingResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "402":
          $ref: "#/components/responses/PaymentRequired"
        "403":
          $ref: "#/components/responses/Forbidden"
  /recurring-booking/precheck:
    post:
      tags: [Recurring Bookings]
      summary: Pre-check recurring booking creation
      description: Validates whether a recurring booking can be created without actually creating it. Returns the validation result for each planned occurrence.
      operationId: preCheckRecurringBooking
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRecurringBookingRequest"
      responses:
        "200":
          description: Validation results for each occurrence
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CreateRecurringBookingResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "402":
          $ref: "#/components/responses/PaymentRequired"
        "403":
          $ref: "#/components/responses/Forbidden"
  /recurring-booking/{id}:
    get:
      tags: [Recurring Bookings]
      summary: Get a recurring booking
      description: Returns a single recurring booking by ID. Users can view their own; Space Admins can view any.
      operationId: getRecurringBooking
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Recurring booking details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GetRecurringBookingResponse"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      tags: [Recurring Bookings]
      summary: Delete a recurring booking
      description: Deletes a recurring booking and all its future individual bookings. Users can delete their own; Space Admins can delete any.
      operationId: deleteRecurringBooking
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "204":
          $ref: "#/components/responses/Updated"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /recurring-booking/{id}/ical:
    get:
      tags: [Recurring Bookings]
      summary: Download recurring booking as iCal
      description: Downloads all bookings of a recurring booking series as an iCalendar (.ics) file.
      operationId: getRecurringBookingIcal
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: iCalendar file
          content:
            text/calendar:
              schema:
                type: string
          headers:
            Content-Disposition:
              schema:
                type: string
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"

  # ===========================
  # Locations
  # ===========================
...